emotion_detection_v2/requirements.txt
# Core ML
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
albumentations==1.3.1

# Training & Logging
wandb==0.16.0
tensorboard==2.15.1
tqdm==4.66.1

# Data & Utils
numpy==1.24.3
pandas==2.1.3
opencv-python==4.8.1.78
pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Inference & Deployment
onnx==1.15.0
onnxruntime==1.16.3
